reorg bridge commands into one file. make parent bridge command for proper help text. add missing fields to proto.
